rename shop app component in PascalCase
