+++ src/ShopApp.tsx
@@ -8,7 +8,7 @@ import { Form } from './components/form'
 import logo from './images/droppe-logo.png'
 import img1 from './images/img1.png'
 import img2 from './images/img2.png'
-import styles from './shopApp.module.css'
+import styles from './ShopApp.module.css'
 import Header from 'components/Header'
 import Banner from 'components/Banner'
 import Stats from 'components/Stats'
