+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import Modal from 'react-modal'
 import './index.css'
-import { ShopApp } from './shop-app'
+import { ShopApp } from './ShopApp'
 import * as serviceWorker from './serviceWorker'
 
 Modal.setAppElement('#root')
